created pages and routes
