+++ music/src/pages/category/index.jsx
@@ -1,0 +1,7 @@
+export const CategoryPage = () => {
+    return (
+      <section>
+        <h1>CategoryPage</h1>
+      </section>
+    );
+  };
