+++ music/src/pages/favorites/index.jsx
@@ -1,0 +1,7 @@
+export const FavoritesPage = () => {
+    return (
+      <div>
+        <h1>FavoritesPage</h1>
+      </div>
+    );
+  };
