+++ music/src/pages/login/index.jsx
@@ -1,0 +1,7 @@
+export const LoginPage = () => {
+    return (
+      <div>
+        <h1>LoginPage</h1>
+      </div>
+    );
+  };
