+++ music/src/pages/not-found/index.jsx
@@ -1,0 +1,7 @@
+export const NotFoundPage = () => {
+    return (
+      <div>
+        <h1>NotFoundPage</h1>
+      </div>
+    );
+  };
