+++ music/src/pages/register/index.jsx
@@ -1,0 +1,7 @@
+export const RegisterPage = () => {
+    return (
+      <div>
+        <h1>RegisterPage</h1>
+      </div>
+    );
+  };
