+++ music/src/routes.jsx
@@ -1,0 +1,33 @@
+import { Routes, Route } from "react-router-dom";
+import { MainPage } from "./pages/main";
+import { CategoryPage } from "./pages/category";
+import { FavoritesPage } from "./pages/favorites";
+import { NotFoundPage } from "./pages/not-found";
+import { LoginPage } from "./pages/login";
+
+export const AppRoutes = () => {
+  return (
+    <Routes>
+      <Route
+        path="/login"
+        element={<LoginPage />}
+      />
+      <Route
+        path="/main"
+        element={<MainPage />}
+      />
+      <Route
+        path="/favorites"
+        element={<FavoritesPage />}
+      />
+      <Route
+        path="/category/:id"
+        element={<CategoryPage />}
+      />
+      <Route
+        path="*"
+        element={<NotFoundPage />}
+      />
+    </Routes>
+  );
+};
